% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo2001.R
\name{censo2001}
\alias{censo2001}
\title{Download data from 2001 Spanish Census from INE}
\usage{
censo2001(file)
}
\arguments{
\item{file}{Character string with the name of the microdata files provided by
the INE on the \href{https://www.ine.es/censos2011_datos/cen11_datos_microdatos.htm#}{Census}
section. The 2001 files have been separated into 6 different \code{txt} files.
\code{file} supports a vector with the path to these 6 files and read them
automatically. It allows either a path to the files, 
or literal data (single string or raw vector). It also allows compressed
files in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} format.}
}
\value{
\code{\link[tibble]{tibble}} with all avaliable requested data where each row corresponds to an anonymised citizen.
}
\description{
This function reads the data from 2001 Spanish Census using the information provided by the Spanish Statistical Office (INE).
}
\details{
This function reads microdata from the 2001 census in Spain. It was originally built using the metadata for 2010 file but it may be used for earlier/later years.

Note that the loaded object will be of considerable size and may require a computer from 8 GB of RAM. In order to avoid this kind of computational issues, you can also download the splitted version provided by the INE.
}
\note{
The \code{file} parameter allows any flat file with fixed width.
}
\examples{

# First download ftp://www.ine.es/temas/censopv/cen01_ph/parte1.zip
# Then ftp://www.ine.es/temas/censopv/cen01_ph/parte2.zip
# ...
# Until ftp://www.ine.es/temas/censopv/cen01_ph/parte6.zip
# Be sure to download from \code{part1} to \code{part6}.

# Suppose that the file path to the 6 text files in your local computer are
# contained in x. Then:
\dontrun{
raw <- censo2001(x)
summary(raw)
}
}
\references{
http://www.ine.es/prodyser/microdatos.htm
}
\seealso{
\code{\link[vroom]{vroom_fwf}} to read fixed width files.
}
\author{
Jorge cimentada adapting the code from Carlos Neira and Carlos J. Gil Bellosta
}
