% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontur2016.R
\name{frontur2016}
\alias{frontur2016}
\title{Download data for "Estadística de movimientos turísticos en frontera - FRONTUR" from INE}
\usage{
frontur2016(file)
}
\arguments{
\item{file}{Character string with the name of the microdata file provided by the INE on the
\href{http://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736176996&menu=resultados&secc=1254736195382&idp=1254735576863}{FRONTUR} section. It allows either a path to a file 
or literal data (single string or raw vector). It also allows compressed files in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} format.}
}
\value{
\code{\link[tibble]{tibble}} with all avaliable requested data where each row corresponds to an anonymised citizen.
}
\description{
This function downloads FRONTUR statistcs from the Spanish Statistical Office (INE).
}
\details{
This function reads deaths microdata in Spain.

It was originally built using the metadata for the 2016 file and may work with microdata files for other yearly submissions. Previous files have a different, non compatible, format.
}
\note{
The \code{file} parameter allows any flat file with fixed width.
}
\examples{
\dontrun{
raw <- frontur2016("datos_1_19.zip") # It will be automatically uncompressed.
summary(raw)
}
}
\references{
http://www.ine.es/prodyser/microdatos.htm
}
\seealso{
\code{\link[read_fwf]{readr}} to read fixed width files.
}
\author{
Carlos J. Gil Bellosta, David Carrasco
}
